+++ yumemicoding/Models/MyProfileStructure.swift
@@ -23,11 +23,7 @@ class YearMonthDay {
 
 extension YearMonthDay {
     func getDate() -> Date {
-        var dateComponents = DateComponents()
-        
-        dateComponents.year = self.year
-        dateComponents.month = self.month
-        dateComponents.day = self.day
+        let dateComponents = DateComponents(year: self.year, month: self.month, day: self.day)
         
         if let customDate = Calendar.current.date(from: dateComponents) {
             return customDate
@@ -38,20 +34,8 @@ extension YearMonthDay {
     }
 }
 
-extension Date {
-    
-    func getYearMonthDay() -> YearMonthDay {
-        let calendar = Calendar(identifier: .gregorian)
-        let year = calendar.component(.year, from: self)
-        let month = calendar.component(.month, from: self)
-        let day = calendar.component(.day, from: self)
-        
-        return YearMonthDay(year: year, month: month, day: day)
-    }
-}
-
 @Model
-class MyStatus: ObservableObject {
+class MyProfile: ObservableObject {
     var name: String
     var birthday: YearMonthDay
     var bloodType: String
